requirements.txt
fastapi
uvicorn
pinecone
sentence-transformers
openai
python-dotenv
httpx
torch
sqlalchemy
pandas
pymysql
